add tx  gaslimit check

+++ src/xtopcom/xdata/xtransaction_v1.h
@@ -108,7 +108,7 @@ class xtransaction_v1_t : public xbase_dataunit_t<xtransaction_v1_t, xdata_type_
     virtual void set_tx_type(uint16_t type) override {m_transaction_type = type;};
     virtual uint16_t get_tx_type() const override {return m_transaction_type;};
     virtual void set_tx_len(uint16_t len) override {m_transaction_len = len;};
-    virtual uint16_t get_tx_len() const override {return m_transaction_len;};
+    virtual uint32_t get_tx_len() const override {return m_transaction_len;};
     virtual void set_tx_version(uint32_t version) override {m_version = version;};
     virtual uint32_t get_tx_version() const override {return m_version;};
     void set_to_ledger_id(uint16_t id) {m_to_ledger_id = id;};
